+++ CHANGELOG.md
@@ -16,6 +16,7 @@ found to have issues, and we are re-working the feature for a future release.
 - Filter when functionality has been disabled. Users that have filters with
 'when' properties defined should delete and recreate the filter. Filter when
 uses the same facility as check subdue for handling time windows.
+- Removed event.Hooks and event.Silenced deprecated fields
 
 ### Changed
 - Assets and checks environments are now merged, with a preference given to the
@@ -37,6 +38,7 @@ uses the same facility as check subdue for handling time windows.
   "sensu" and "sensu-devel", could be listed together.
 - Fixed a bug in the agent where the agent would deadlock after a significant
 period of disconnection from the backend.
+- Fixed a bug where logging events without checks would cause a nil panic.
 
 ## [2.0.0-beta.8-1] - 2018-11-15